chore: refactor image file name and folder structure
