+++ src/components/custom/_signup-public.tsx
@@ -10,10 +10,10 @@ import { Flex, Box, Container, Desktop, Mobile } from 'components/containers'
 import { deriv_app_url } from 'common/constants'
 import device from 'themes/device'
 // SVG
-import Apple from 'images/svg/custom/apple.svg'
+import Apple from 'images/svg/components/component_apple.svg'
 import Facebook from 'images/svg/custom/facebook-blue.svg'
 import Google from 'images/svg/custom/google.svg'
-import Arrow from 'images/svg/custom/chevron-right.svg'
+import Arrow from 'images/svg/components/component_chevron_right.svg'
 import { useIsRtl } from 'components/hooks/use-isrtl'
 
 type SocialButtonContent = {
