+++ src/components/custom/other-platforms.tsx
@@ -8,11 +8,11 @@ import { TString } from 'types/generics'
 import { binary_bot_url } from 'common/constants'
 import device from 'themes/device'
 // icons
-import Blog from 'images/svg/custom/blog-nav.svg'
+import Blog from 'images/svg/components/component_blog_nav.svg'
 import Career from 'images/svg/menu/careers.svg'
 import Choose from 'images/svg/menu/choose.svg'
-import Cryptocurrencies from 'images/svg/custom/cryptocurrencies-nav.svg'
-import Commodities from 'images/svg/custom/commodities-nav.svg'
+import Cryptocurrencies from 'images/svg/components/component_cryptocurrencies_nav.svg'
+import Commodities from 'images/svg/components/component_commodities_nav.svg'
 import Community from 'images/svg/menu/community.svg'
 import Contact from 'images/svg/menu/contact.svg'
 import DerivedFX from 'images/svg/custom/derived-fx.svg'
