+++ src/components/elements/accordion.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react'
 import styled from 'styled-components'
 import { Text } from './typography'
-import Chevron from 'images/svg/custom/chevron-bottom.svg'
-import ChevronThick from 'images/svg/custom/chevron-thick.svg'
+import Chevron from 'images/svg/components/component_chevron_bottom.svg'
+import ChevronThick from 'images/svg/chevron_thick.svg'
 import Minus from 'images/svg/elements/minus.svg'
 import Plus from 'images/svg/elements/plus.svg'
 import device from 'themes/device'
