+++ src/components/elements/off-canvas-menu.tsx
@@ -17,10 +17,10 @@ import {
 import { binary_bot_url, deriv_status_page_url } from 'common/constants'
 // SVG
 import AffiliateIb from 'images/svg/menu/affiliate-ib.svg'
-import Blog from 'images/svg/custom/blog-nav.svg'
+import Blog from 'images/svg/components/component_blog_nav.svg'
 import Career from 'images/svg/menu/careers.svg'
 import Choose from 'images/svg/menu/choose.svg'
-import Commodities from 'images/svg/custom/commodities-nav.svg'
+import Commodities from 'images/svg/components/component_commodities_nav.svg'
 import Community from 'images/svg/menu/community.svg'
 import Contact from 'images/svg/menu/contact.svg'
 import DerivedFX from 'images/svg/custom/derived-fx.svg'
@@ -48,7 +48,7 @@ import SecureTrading from 'images/svg/menu/secure-trading.svg'
 import Smarttrader from 'images/svg/custom/smarttrader.svg'
 import Status from 'images/svg/elements/status.svg'
 import StockIndices from 'images/svg/custom/stock-indices-nav.svg'
-import Cryptocurrencies from 'images/svg/custom/cryptocurrencies-nav.svg'
+import Cryptocurrencies from 'images/svg/components/component_cryptocurrencies_nav.svg'
 import Story from 'images/svg/menu/story.svg'
 import Terms from 'images/svg/menu/terms.svg'
 import Trade from 'images/svg/custom/trader-tool-nav.svg'
