+++ src/pages/contact_us/_offices.tsx
@@ -1,26 +1,26 @@
 import React from 'react'
 import { graphql, useStaticQuery } from 'gatsby'
 import styled from 'styled-components'
+import France from 'images/svg/contact-us/contact_us_france.svg'
+import Malta from 'images/svg/contact-us/contact_us_malta.svg'
+import Guernsey from 'images/svg/contact-us/contact_us_guernsey.svg'
+import Dubai from 'images/svg/contact-us/contact_us_dubai.svg'
+import Paraguay from 'images/svg/contact-us/contact_us_paraguay.svg'
+import Malaysia from 'images/svg/contact-us/contact_us_malaysia.svg'
+import Cayman from 'images/svg/contact-us/contact_us_cayman.svg'
+import Cyprus from 'images/svg/contact-us/contact_us_cyprus.svg'
+import Rwanda from 'images/svg/contact-us/contact_us_rwanda.svg'
+import Singapore from 'images/svg/contact-us/contact_us_singapore.svg'
+import Jordan from 'images/svg/contact-us/contact_us_jordan.svg'
+import Vanuatu from 'images/svg/contact-us/contact_us_vanuatu.svg'
+import UK from 'images/svg/contact-us/contact_us_uk.svg'
+import Berlin from 'images/svg/contact-us/contact_us_berlin.svg'
+import Bvi from 'images/svg/contact-us/contact_us_bvi.svg'
+import HongKong from 'images/svg/contact-us/contact_us_hongkong.svg'
 import { Header, QueryImage, Text, LocalizedLinkText } from 'components/elements'
 import { SectionContainer, Container, Flex, Box, Desktop, Mobile } from 'components/containers'
 import { Localize, LocalizedLink } from 'components/localization'
 import device from 'themes/device'
-import France from 'images/svg/contact/france.svg'
-import Malta from 'images/svg/contact/malta.svg'
-import Guernsey from 'images/svg/contact/guernsey.svg'
-import Dubai from 'images/svg/contact/dubai.svg'
-import Paraguay from 'images/svg/contact/paraguay.svg'
-import Malaysia from 'images/svg/contact/malaysia.svg'
-import Cayman from 'images/svg/contact/cayman.svg'
-import Cyprus from 'images/svg/contact/cyprus.svg'
-import Rwanda from 'images/svg/contact/rwanda.svg'
-import Singapore from 'images/svg/contact/singapore.svg'
-import Jordan from 'images/svg/contact/jordan.svg'
-import Vanuatu from 'images/svg/contact/vanuatu.svg'
-import UK from 'images/svg/contact/uk.svg'
-import Berlin from 'images/svg/contact/berlin.svg'
-import Bvi from 'images/svg/contact/bvi.svg'
-import HongKong from 'images/svg/contact/hongkong.svg'
 
 type GridLayoutProps = {
     children?: React.ReactNode | React.ReactChild
