+++ src/pages/landing/ebooks/components/_get-ebook.tsx
@@ -10,7 +10,7 @@ import { Localize, localize } from 'components/localization'
 import { Flex } from 'components/containers'
 import AgreementLabel from 'components/custom/_agreement-label'
 import device from 'themes/device'
-import Apple from 'images/svg/custom/apple.svg'
+import Apple from 'images/svg/components/component_apple.svg'
 import Facebook from 'images/svg/custom/facebook-blue.svg'
 import Google from 'images/svg/custom/google.svg'
 import ViewEmailImage from 'images/common/sign-up/view-email.png'
